// src/common/helper/enums.ts
export enum DeviceStatus {
    ONLINE = 'online',
    OFFLINE = 'offline',
}

export enum BatteryHealth {
    GOOD = 'good',
    MODERATE = 'moderate',
    CRITICAL = 'critical',
    DEAD = 'dead',
}


export enum CustomerType {
    INFORMATION = 'INFORMATION',
    DEVICE = 'DEVICE',
    LOGS = 'LOGS',
}